yik yak liq token

+++ interfaces/IYakStrategy.sol
@@ -1,26 +1,30 @@
-// SPDX-License-Identifier: BUSL-1.1
 pragma solidity ^0.8.0;
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 interface IYakStrategy {
-
     /**
      * @notice Deposit and deploy deposits tokens to the strategy
      * @dev Must mint receipt tokens to `msg.sender`
      * @param amount deposit tokens
      */
-    function deposit(uint amount) external;
+    function deposit(uint256 amount) external;
 
     /**
-    * @notice Deposit using Permit
-    * @dev Should revert for tokens without Permit
-    * @param amount Amount of tokens to deposit
-    * @param deadline The time at which to expire the signature
-    * @param v The recovery byte of the signature
-    * @param r Half of the ECDSA signature pair
-    * @param s Half of the ECDSA signature pair
-    */
-    function depositWithPermit(uint amount, uint deadline, uint8 v, bytes32 r, bytes32 s) external;
+     * @notice Deposit using Permit
+     * @dev Should revert for tokens without Permit
+     * @param amount Amount of tokens to deposit
+     * @param deadline The time at which to expire the signature
+     * @param v The recovery byte of the signature
+     * @param r Half of the ECDSA signature pair
+     * @param s Half of the ECDSA signature pair
+     */
+    function depositWithPermit(
+        uint256 amount,
+        uint256 deadline,
+        uint8 v,
+        bytes32 r,
+        bytes32 s
+    ) external;
 
     /**
      * @notice Deposit on behalf of another account
@@ -28,13 +32,13 @@ interface IYakStrategy {
      * @param account address to receive receipt tokens
      * @param amount deposit tokens
      */
-    function depositFor(address account, uint amount) external;
+    function depositFor(address account, uint256 amount) external;
 
     /**
      * @notice Redeem receipt tokens for deposit tokens
      * @param amount receipt tokens
      */
-    function withdraw(uint amount) external;
+    function withdraw(uint256 amount) external;
 
     /**
      * @notice Reinvest reward tokens into deposit tokens
@@ -45,26 +49,29 @@ interface IYakStrategy {
      * @notice Estimate reinvest reward
      * @return reward tokens
      */
-    function estimateReinvestReward() external view returns (uint);
+    function estimateReinvestReward() external view returns (uint256);
 
     /**
      * @notice Reward tokens avialable to strategy, including balance
      * @return reward tokens
      */
-    function checkReward() external view returns (uint);
+    function checkReward() external view returns (uint256);
 
     /**
      * @notice Estimated deposit token balance deployed by strategy, excluding balance
      * @return deposit tokens
      */
-    function estimateDeployedBalance() external view returns (uint);
+    function estimateDeployedBalance() external view returns (uint256);
 
     /**
      * @notice Rescue all available deployed deposit tokens back to Strategy
      * @param minReturnAmountAccepted min deposit tokens to receive
      * @param disableDeposits bool
      */
-    function rescueDeployedFunds(uint minReturnAmountAccepted, bool disableDeposits) external;
+    function rescueDeployedFunds(
+        uint256 minReturnAmountAccepted,
+        bool disableDeposits
+    ) external;
 
     /**
      * @notice Calculate receipt tokens for a given amount of deposit tokens
@@ -73,12 +80,18 @@ interface IYakStrategy {
      * @param amount deposit tokens
      * @return receipt tokens
      */
-    function getSharesForDepositTokens(uint amount) external view returns (uint);
+    function getSharesForDepositTokens(uint256 amount)
+        external
+        view
+        returns (uint256);
 
     /**
      * @notice Calculate deposit tokens for a given amount of receipt tokens
      * @param amount receipt tokens
      * @return deposit tokens
      */
-    function getDepositTokensForShares(uint amount) external view returns (uint);
+    function getDepositTokensForShares(uint256 amount)
+        external
+        view
+        returns (uint256);
 }
